Renamed implementation classes to match unity version
